+++ bootstrap/src/test/java/onlydust/com/marketplace/api/bootstrap/it/api/MeApiIT.java
@@ -13,17 +13,16 @@ import onlydust.com.marketplace.api.bootstrap.helper.Auth0ApiClientStub;
 import onlydust.com.marketplace.api.bootstrap.helper.CurrencyHelper;
 import onlydust.com.marketplace.api.bootstrap.helper.UserAuthHelper;
 import onlydust.com.marketplace.api.bootstrap.suites.tags.TagMe;
-import onlydust.com.marketplace.api.contract.model.*;
-import onlydust.com.marketplace.api.postgres.adapter.entity.write.old.ApplicationEntity;
+import onlydust.com.marketplace.api.contract.model.RewardItemRequest;
+import onlydust.com.marketplace.api.contract.model.RewardRequest;
+import onlydust.com.marketplace.api.contract.model.RewardType;
 import onlydust.com.marketplace.api.postgres.adapter.entity.write.old.ProjectLeadEntity;
 import onlydust.com.marketplace.api.postgres.adapter.entity.write.old.ProjectLeaderInvitationEntity;
 import onlydust.com.marketplace.api.postgres.adapter.entity.write.old.ProjectRepoEntity;
-import onlydust.com.marketplace.api.postgres.adapter.repository.old.ApplicationRepository;
 import onlydust.com.marketplace.api.postgres.adapter.repository.old.ProjectLeadRepository;
 import onlydust.com.marketplace.api.postgres.adapter.repository.old.ProjectLeaderInvitationRepository;
 import onlydust.com.marketplace.api.postgres.adapter.repository.old.ProjectRepoRepository;
 import onlydust.com.marketplace.api.rest.api.adapter.mapper.DateMapper;
-import onlydust.com.marketplace.project.domain.model.Application;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
@@ -47,8 +46,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class MeApiIT extends AbstractMarketplaceApiIT {
     @Autowired
     ProjectLeaderInvitationRepository projectLeaderInvitationRepository;
-    @Autowired
-    ApplicationRepository applicationRepository;
     @Autowired
     ProjectLeadRepository projectLeadRepository;
     @Autowired
@@ -192,168 +189,6 @@ public class MeApiIT extends AbstractMarketplaceApiIT {
                 .jsonPath(format("$.leaders[?(@.githubUserId==%d)]", githubUserId)).doesNotExist();
     }
 
-    @SneakyThrows
-    @Test
-    void should_apply_to_project() {
-        // Given
-        final var user = userAuthHelper.authenticateAnthony();
-        final var issueId = 1974127467L;
-        final var motivations = faker.lorem().paragraph();
-        final var problemSolvingApproach = faker.lorem().paragraph();
-        final var projectId = UUID.fromString("7d04163c-4187-4313-8066-61504d34fc56");
-
-        final var request = new ApplicationRequest()
-                .projectId(projectId)
-                .issueId(issueId)
-                .motivation(motivations)
-                .problemSolvingApproach(problemSolvingApproach);
-
-        githubWireMockServer.stubFor(post(urlEqualTo("/repository/380954304/issues/7/comments"))
-                .willReturn(aResponse()
-                        .withStatus(200)
-                        .withBody("""
-                                {
-                                    "id": 123456789
-                                }
-                                """)));
-
-        // When
-        final var applicationId = client.post()
-                .uri(getApiURI(ME_APPLICATIONS))
-                .header("Authorization", BEARER_PREFIX + user.jwt())
-                .contentType(MediaType.APPLICATION_JSON)
-                .bodyValue(request)
-                // Then
-                .exchange()
-                .expectStatus()
-                .isOk()
-                .expectBody(ApplicationResponse.class)
-                .returnResult().getResponseBody().getId();
-
-        final var application = applicationRepository.findById(applicationId).orElseThrow();
-        assertThat(application.projectId()).isEqualTo(projectId);
-        assertThat(application.applicantId()).isEqualTo(user.user().getGithubUserId());
-        assertThat(application.issueId()).isEqualTo(issueId);
-        assertThat(application.origin()).isEqualTo(Application.Origin.MARKETPLACE);
-        assertThat(application.commentId()).isEqualTo(123456789L);
-        assertThat(application.motivations()).isEqualTo(motivations);
-        assertThat(application.problemSolvingApproach()).isEqualTo(problemSolvingApproach);
-    }
-
-    @Test
-    void should_not_be_able_to_apply_twice() {
-        // Given
-        final var user = userAuthHelper.authenticateAnthony();
-        final var issueId = 1736474921L;
-        final var projectId = UUID.fromString("7d04163c-4187-4313-8066-61504d34fc56");
-
-        applicationRepository.save(new ApplicationEntity(
-                UUID.randomUUID(),
-                ZonedDateTime.now(),
-                projectId,
-                user.user().getGithubUserId(),
-                Application.Origin.MARKETPLACE,
-                issueId,
-                111L,
-                "My motivations",
-                null
-        ));
-
-        final var motivations = faker.lorem().paragraph();
-        final var problemSolvingApproach = faker.lorem().paragraph();
-
-        final var request = new ApplicationRequest()
-                .projectId(projectId)
-                .issueId(issueId)
-                .motivation(motivations)
-                .problemSolvingApproach(problemSolvingApproach);
-
-        // When
-        client.post()
-                .uri(getApiURI(ME_APPLICATIONS))
-                .header("Authorization", BEARER_PREFIX + user.jwt())
-                .contentType(MediaType.APPLICATION_JSON)
-                .bodyValue(request)
-                // Then
-                .exchange()
-                .expectStatus()
-                .isBadRequest();
-    }
-
-    @Test
-    void should_update_application() {
-        // Given
-        final var user = userAuthHelper.authenticateAnthony();
-        final var applicationId = UUID.randomUUID();
-
-        applicationRepository.save(new ApplicationEntity(
-                applicationId,
-                ZonedDateTime.now(),
-                UUID.fromString("7d04163c-4187-4313-8066-61504d34fc56"),
-                user.user().getGithubUserId(),
-                Application.Origin.GITHUB,
-                1974137199L,
-                111L,
-                "My motivations",
-                null
-        ));
-
-        final var motivations = faker.lorem().paragraph();
-        final var problemSolvingApproach = faker.lorem().paragraph();
-
-        final var request = new ApplicationUpdateRequest()
-                .motivation(motivations)
-                .problemSolvingApproach(problemSolvingApproach);
-
-        // When
-        client.put()
-                .uri(getApiURI(ME_APPLICATION.formatted(applicationId)))
-                .header("Authorization", BEARER_PREFIX + user.jwt())
-                .contentType(MediaType.APPLICATION_JSON)
-                .bodyValue(request)
-                // Then
-                .exchange()
-                .expectStatus()
-                .isNoContent();
-
-        final var application = applicationRepository.findById(applicationId).orElseThrow();
-        assertThat(application.origin()).isEqualTo(Application.Origin.MARKETPLACE);
-        assertThat(application.motivations()).isEqualTo(motivations);
-        assertThat(application.problemSolvingApproach()).isEqualTo(problemSolvingApproach);
-    }
-
-    @Test
-    void should_not_be_able_to_apply_to_non_existing_project() {
-        // Given
-        final var githubUserId = faker.number().randomNumber();
-        final var login = faker.name().username();
-        final var avatarUrl = faker.internet().avatar();
-        final var userId = UUID.randomUUID();
-        final String jwt = userAuthHelper.newFakeUser(userId, githubUserId, login, avatarUrl, false).jwt();
-
-        final var projectId = UUID.fromString("77777777-4444-4444-4444-61504d34fc56");
-        final var issueId = 1736474921L;
-        final var motivations = faker.lorem().paragraph();
-        final var problemSolvingApproach = faker.lorem().paragraph();
-
-        final var request = new ApplicationRequest()
-                .projectId(projectId)
-                .issueId(issueId)
-                .motivation(motivations)
-                .problemSolvingApproach(problemSolvingApproach);
-
-        // When
-        client.post()
-                .uri(getApiURI(ME_APPLICATIONS))
-                .header("Authorization", BEARER_PREFIX + jwt)
-                .contentType(MediaType.APPLICATION_JSON)
-                .bodyValue(request)
-                // Then
-                .exchange()
-                .expectStatus()
-                .isNotFound();
-    }
-
     @Test
     void should_return_projects_led_and_applications() {
         // Given
@@ -368,31 +203,6 @@ public class MeApiIT extends AbstractMarketplaceApiIT {
 
         final var projectAppliedTo1 = UUID.fromString("dcb3548a-977a-480e-8fb4-423d3f890c04");
         final var projectAppliedTo2 = UUID.fromString("c66b929a-664d-40b9-96c4-90d3efd32a3c");
-
-        applicationRepository.saveAll(List.of(
-                new ApplicationEntity(
-                        UUID.randomUUID(),
-                        ZonedDateTime.now(),
-                        projectAppliedTo1,
-                        pierre.user().getGithubUserId(),
-                        Application.Origin.MARKETPLACE,
-                        1736474921L,
-                        112L,
-                        "My motivations",
-                        null
-                ),
-                new ApplicationEntity(
-                        UUID.randomUUID(),
-                        ZonedDateTime.now(),
-                        projectAppliedTo2,
-                        pierre.user().getGithubUserId(),
-                        Application.Origin.MARKETPLACE,
-                        1736504583L,
-                        113L,
-                        "My motivations",
-                        null
-                )
-        ));
 
         // When
         client.get()
@@ -417,10 +227,7 @@ public class MeApiIT extends AbstractMarketplaceApiIT {
                                                                 ".com/5003677688814069549.png")
                 .jsonPath("$.createdAt").isEqualTo(DateMapper.toZoneDateTime(pierre.user().getCreatedAt()).format(DateTimeFormatter.ISO_INSTANT))
                 .jsonPath("$.projectsLed[0].slug").isEqualTo("bretzel")
-                .jsonPath("$.pendingProjectsLed.length()").isEqualTo(0)
-                .jsonPath("$.projectsAppliedTo.length()").isEqualTo(2)
-                .jsonPath("$.projectsAppliedTo[0]").isEqualTo(projectAppliedTo1.toString())
-                .jsonPath("$.projectsAppliedTo[1]").isEqualTo(projectAppliedTo2.toString());
+                .jsonPath("$.pendingProjectsLed.length()").isEqualTo(0);
 
         // Public repo
         projectRepoRepository.save(new ProjectRepoEntity(projectIdWithoutRepo, 593218280L));
@@ -441,10 +248,7 @@ public class MeApiIT extends AbstractMarketplaceApiIT {
                                                                        ".com/cZCHED10RzuEloOXuk7A")
                 .jsonPath("$.pendingProjectsLed[0].slug").isEqualTo("red-bull")
                 .jsonPath("$.pendingProjectsLed[0].hasMissingGithubAppInstallation").isEqualTo(false)
-                .jsonPath("$.pendingProjectsLed.length()").isEqualTo(1)
-                .jsonPath("$.projectsAppliedTo.length()").isEqualTo(2)
-                .jsonPath("$.projectsAppliedTo[0]").isEqualTo(projectAppliedTo1.toString())
-                .jsonPath("$.projectsAppliedTo[1]").isEqualTo(projectAppliedTo2.toString());
+                .jsonPath("$.pendingProjectsLed.length()").isEqualTo(1);
     }
 
     @Test
