+++ infrastructure/postgres-adapter/src/main/java/onlydust/com/marketplace/api/postgres/adapter/PostgresProjectAdapter.java
@@ -414,4 +414,12 @@ public class PostgresProjectAdapter implements ProjectStoragePort {
     public List<UUID> getProjectContributedOnIdsForUser(UUID userId) {
         return projectRepository.getProjectContributedOnIdsForUser(userId);
     }
+
+    @Override
+    public List<UUID> findProjectIdsByRepoId(Long repoId) {
+        return projectRepoRepository.findAllByRepoId(repoId)
+                .stream()
+                .map(ProjectRepoEntity::getProjectId)
+                .toList();
+    }
 }
