+++ infrastructure/postgres-adapter/src/main/java/onlydust/com/marketplace/api/postgres/adapter/repository/old/ProjectRepoRepository.java
@@ -4,7 +4,11 @@ import onlydust.com.marketplace.api.postgres.adapter.entity.write.old.ProjectRep
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.util.List;
+
 public interface ProjectRepoRepository extends JpaRepository<ProjectRepoEntity, ProjectRepoEntity.PrimaryKey> {
     @Query("SELECT EXISTS (SELECT 1 FROM ProjectRepoEntity p WHERE p.repoId = :repoId)")
     boolean existsByRepoId(Long repoId);
+
+    List<ProjectRepoEntity> findAllByRepoId(Long repoId);
 }
